This patch adds the possibility to add backlash to the axis. You may add following in the config file : x_axis_backlash                              0.05             # Backlash of X-axis y_axis_backlash                              0.05             # Backlash of Y-axis z_axis_backlash                              0.0              # Backlash of Z-axis
Additional distance to move is calculated in Robot depending on direction change (last direction is stored in axis_direction. Extruder still needs the old distance to travel (distance in block and robot) and the new calculated moving millimeters is stored in millimeters. PLanner needs to calculate with millimeters(way + backlash), Extruder needs to calculate with distance (original way to move).